fix: RSS-TD-02 solve code review comments, refactor code for router
